// src/app/sites/[site]/[theme]/static/[...slug]/queries.generated.ts
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
const defaultOptions = {} as const;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** country code */
  CountryCode: any;
  /** An RFC-3339 compliant Full Date Scalar */
  Date: any;
  /** An RFC-3339 compliant DateTime Scalar */
  DateTime: any;
  /** A field whose value conforms to the standard internet email address format as specified in RFC822: https://www.w3.org/Protocols/rfc822/. */
  EmailAddress: any;
  /** gender */
  Gender: any;
  /** Long type */
  Long: any;
};
// ...
export type StaticPageFragment = { __typename: 'BPD_StaticPage', id: string, template: string, content?: string | null, title?: string | null, subTemplate?: string | null, menu?: Array<{ id: string, name?: string | null, uri: { __typename: 'BPD_ExternalUri', url?: string | null } | { __typename: 'BPD_Uri', url?: string | null } }> | null };

export type AppRouterStaticPageQueryVariables = Types.Exact<{
  host: Types.Scalars['String'];
  pathname: Types.Scalars['String'];
}>;


export type AppRouterStaticPageQuery = { page?: { __typename: 'BPD_ListPage', id: string, template: string } | { __typename: 'BPD_NonPage', id: string, template: string } | { __typename: 'BPD_ProductPage', id: string, template: string } | { __typename: 'BPD_RegionPage', id: string, template: string } | { __typename: 'BPD_StaticPage', id: string, template: string, content?: string | null, title?: string | null, subTemplate?: string | null, menu?: Array<{ id: string, name?: string | null, uri: { __typename: 'BPD_ExternalUri', url?: string | null } | { __typename: 'BPD_Uri', url?: string | null } }> | null } | null };

export const StaticPageFragmentDoc = gql`
    fragment StaticPage on BPD_StaticPage {
  id
  __typename
  template
  content
  title
  subTemplate
  menu {
    id
    name
    uri {
      __typename
      ... on BPD_ExternalUri {
        url
      }
      ... on BPD_Uri {
        url
      }
    }
  }
}
    `;
export const AppRouterStaticPageDocument = gql`
    query AppRouterStaticPage($host: String!, $pathname: String!) {
  page: bpd_page(host: $host, pathname: $pathname) {
    id
    __typename
    template
    ...StaticPage
  }
}
    ${StaticPageFragmentDoc}`;